Renamed CargoItemList into EmptyItemList
Now it can be used anywhere
