+++ UnityProject/Assets/Scripts/UI/FilledImageAnimation.cs
